# Conflicts:
#	Gateway/Request/CheckoutDataBuilder.php

+++ Api/Data/CreditmemoInterface.php
@@ -83,15 +83,15 @@ interface CreditmemoInterface
     public function setCreditmemoId(int $creditmemoId): CreditmemoInterface;
 
     /**
-     * @return int|null
+     * @return float|null
      */
-    public function getAmount(): ?int;
+    public function getAmount(): ?float;
 
     /**
-     * @param int $amount
+     * @param float $amount
      * @return $this
      */
-    public function setAmount(int $amount): CreditmemoInterface;
+    public function setAmount(float $amount): CreditmemoInterface;
 
     /**
      * @return int|null
